// src/generated/models/transition-validation-model.ts
import type { StatusTransition } from './status-transition-model';
import type { Workspace } from './workspace-model';

export const TransitionValidationStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type TransitionValidationStatusKey = keyof typeof TransitionValidationStatusKeyToLabel;

export const TransitionValidationValidationTypeKeyToLabel = {
  'FieldRequired': 'Field Required',
  'ExpectedValue': 'Expected Value',
  'CompletedTasks': 'Completed Tasks',
  'ApprovalsComplete': 'Approvals Complete',
  'ResolutionRequired': 'Resolution Required',
  'MandatoryComment': 'Mandatory Comment'
} as const;
export type TransitionValidationValidationTypeKey = keyof typeof TransitionValidationValidationTypeKeyToLabel;

export interface TransitionValidation {
  /**
   * @displayName Transition Validation ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Error Message
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  errorMessage: string;
  /**
   * @displayName Expected Value
   */
  expectedValue?: string;
  /**
   * @displayName Field Code
   */
  fieldCode?: string;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: TransitionValidationStatusKey;
  /**
   * @displayName Status Transition ID
   * @validationRule Required for create/update operations
   */
  statusTransitionID: Pick<StatusTransition, 'id' | 'transitionName'>;
  /**
   * @displayName Validation Type
   * @validationRule Required for create/update operations
   */
  validationTypeKey: TransitionValidationValidationTypeKey;
  /**
   * @displayName Workspace ID
   * @validationRule Required for create/update operations
   */
  workspaceID: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _TransitionValidation = 'TransitionValidation' as const;